Data from the ASQ-3 study from CADI-UMAG

Data from the ASQ-3 study from CADI-UMAG

Usage

dataset

Format

A data frame with 1266 rows y 53 variables:

profesional_id: Factor. Identifying number of the assessing professional.

profesional_especialidad: Factor. Profession of the professional evaluator.

paciente_id: Factor. Patient identifier number.

sexo_paciente: Factor. Patient's sex.

respondedor_vinculo: Factor. Type of relationship between the respondent and the patient.

evaluacion_tipo: Factor. Type of assessment.

fecha_evaluacion: Date. Date of patient assessment.

fecha_nacimiento: Date. Patient's date of birth.

edad_cronologica_meses: Numeric. Chronological age of the patient in months.

semanas_prematurez: Numeric. Weeks of prematurity.

edad_corregida_meses: Numeric. Chronological age of the patient in months corrected for weeks of prematurity.

diagnostico: Character. Patient's diagnostic group.

asq3_meses: Numeric. Age in months of patient development according to ASQ 3.

comunicacion_q1 - comunicacion_q6: Numeric. Scoring of the individual questions of the communication item.

comunicacion_total: Numeric. Total score of the communication item.

comunicacion_interpretacion: Factor. Interpretation of the communication item.

motora_gruesa_q1 - motora_gruesa_q6: Numeric. Scoring of the individual questions of the gross motor skills item.

motora_gruesa_total: Numeric. Total score of the gross motor skills item.

motora_gruesa_interpretacion: Factor. Interpretation of the gross motor skills item.

motora_fina_q1 - motora_fina_q6: Numeric. Scoring of the individual questions of the fine motor skills item.

motora_fina_total: Numeric. Total score of the fine motor skills item.

motora_fina_interpretacion: Factor. Interpretation of the fine motor skills item.

resolucion_problemas_q1 - resolucion_problemas_q6: Numeric. Scoring of the individual questions of the problem solving item.

resolucion_problemas_total: Numeric. Total score of the problem solving item.

resolucion_problemas_interpretacion: Factor. Interpretation of the problem solving item.

socio_individual_q1 - socio_individual_q6: Numeric. Scoring of the individual questions of the socio-individual skills item.

socio_individual_total: Numeric. Total score of the socio-individual skills item.

socio_individual_interpretacion: Factor. Interpretation of the socio-individual skills item.

Source

CADI-UMAG, Punta Arenas, Chile.

Ages and Stages Questionnaire

The Ages & Stages Questionnaires, Third Edition (ASQ-3) is a developmental screening tool that pinpoints developmental progress in children between the ages of one month to 5.5 years. For more information please see the homepage of the questionnaire ASQ-3.

See also